create an AST
The grammar now creates an which can be converted to
Javascript or a string view. This makes the grammar
less verbose and moves nearly all the logic into helper
class files.

Helper class files define a class and export a creator
function. The creator function is a constructor function
to reduce the amount of new keywords. This makes the
grammar even cleaner.