seperate parser from transpiler
The new version now has a parser, that is build from the grammar
and a seperate transpiler. The parser creates the abtract syntax
tree. The transpiler creates Javascript source code from the AST.

The toJS function was removed from the AST creator classes. Instead
the util/transpile funtion composes a the transpilers created with
the util/createTranspiler function.

A index.js was created in the helper folder to combine the parser
with the transpiler.