implement new transpiler design
transpilers now are implemented as pure functions that take a tree, a
transpile function for recursive calls and an object of plugins.

The tree is a subtree that the transpiler should transpile. The top node
is always of the type the transpiler function is written for. The subnodes
can be either transpiled by the function itself or by recursive calls on
transpile.

Plugins provide state and context to the functions. For example plugins
can keep track of variable scope, imports and the environment.