fix procedures registering parameters in scope
The ScopePlugin now supports an array of ignored variable names, that will
not be registed and also looks in the upper scopes for variables of the
name. This effectively creates a closure. While SetlX does not have
closures without explicit declaration this is needed to access functions
in the upper scope.

This might lead to bugs, but for now there is no way of seperating
variables that are declared as functions from variables that are simply
values.